requirements.txt
speechrecognition
gtts
pyttsx3
pyaudio
torch
transformers
pyttsx3
SpeechRecognition
google-generativeai
requests
pyserial
luma.oled
python-dotenv
pandas
Pillow
llama-cpp-python
python-dotenv

